LayoutTests/media/video-cuepoint-add-expected.txt

TEST(video.currentTime.toFixed(1) == 0.2) OK
TEST(!video.paused) OK
TEST(video.currentTime.toFixed(1) == 0.4) OK
TEST(!video.paused) OK
TEST(video.currentTime.toFixed(1) == 0) OK
TEST(!video.paused) OK
END OF TEST
